+++ packages/core/src/helpers/css-reset.ts
@@ -2,9 +2,9 @@ import { injectGlobal } from '@chakra-ui/vue-system'
 import { cssResetStyles } from '@chakra-ui/c-reset'
 import { ThemeOverride } from '../extend-theme'
 import { get, runIfFn } from '@chakra-ui/utils'
-import { ColorMode } from '@chakra-ui/c-color-mode'
 import { css } from '@chakra-ui/styled-system'
-import { Ref } from 'vue'
+import { ColorModeRef } from '@chakra-ui/c-color-mode'
+import { computed, ref, watch } from 'vue'
 
 /** Injects CSS reset styles */
 export function injectResetStyles() {
@@ -14,13 +14,24 @@ export function injectResetStyles() {
 /** Injects styles from `theme.styles.global` property */
 export function injectThemeGlobalStyles(
   theme: ThemeOverride,
-  colorMode: Ref<ColorMode>
+  colorMode: ColorModeRef
 ) {
   const styleObjectOrFn = get(theme, 'styles.global')
-  const globalStyles = runIfFn(styleObjectOrFn, {
-    theme,
-    colorMode: colorMode.value,
-  })
-  if (!globalStyles) return undefined
-  const styles = css(globalStyles)(theme)
+  const globalStyles = computed(() =>
+    runIfFn(styleObjectOrFn, {
+      theme,
+      colorMode: colorMode.value,
+    })
+  )
+  if (!globalStyles.value) return undefined
+  const id = ref(0)
+  watch(
+    colorMode,
+    () => {
+      const styles = css(globalStyles.value)(theme)
+    },
+    {
+      immediate: true,
+    }
+  )
 }