+++ packages/core/src/helpers/plugin.types.ts
@@ -1,0 +1,20 @@
+import { ColorModeRef } from '@chakra-ui/c-color-mode'
+import { UnwrapRef } from 'vue-demi'
+import { ThemeOverride } from '../extend-theme'
+
+interface ExtendIconsPath {
+  path: string
+  viewBox?: string
+}
+
+interface IconsOptions {
+  pack?: 'fa' | 'fe'
+  library?: {}
+  extend?: Record<string, ExtendIconsPath>
+}
+export interface ChakraPluginOptions {
+  cssReset?: boolean
+  extendTheme?: ThemeOverride
+  icons?: IconsOptions
+  defaultColorMode?: UnwrapRef<ColorModeRef>
+}