+++ packages/c-portal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chakra-ui/c-portal
 
+## 0.1.0-alpha.5
+
+### Minor Changes
+
+- [#83](https://github.com/chakra-ui/chakra-ui-vue-next/pull/83) [`ac28d42`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/ac28d4217c3afb83383801316876f982163c5f5c) Thanks [@codebender828](https://github.com/codebender828)! - Create breadcrumb and collapse components
+
 ## 0.1.0-alpha.4
 
 ### Patch Changes