+++ packages/c-spinner/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chakra-ui/c-spinner
 
+## 1.0.0-alpha.5
+
+### Minor Changes
+
+- [#83](https://github.com/chakra-ui/chakra-ui-vue-next/pull/83) [`ac28d42`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/ac28d4217c3afb83383801316876f982163c5f5c) Thanks [@codebender828](https://github.com/codebender828)! - Create breadcrumb and collapse components
+
+### Patch Changes
+
+- Updated dependencies [[`ac28d42`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/ac28d4217c3afb83383801316876f982163c5f5c)]:
+  - @chakra-ui/c-visually-hidden@1.0.0-alpha.5
+  - @chakra-ui/vue-system@0.1.0-alpha.5
+  - @chakra-ui/vue-theme@0.1.0-alpha.5
+
 ## 1.0.0-alpha.4
 
 ### Patch Changes